results/plaid-isolated/sonnet-low-run7-files/requirements.txt
plaid-python==25.1.0
flask==3.1.0
python-dotenv==1.0.1
PyJWT==2.10.1
cryptography==44.0.2
